MIgrate notification srtate to separate component
