+++ src/features/parser/actions.js
@@ -5,8 +5,6 @@ export const ADD_WARRIOR = 'parser/ADD_WARRIOR'
 export const REMOVE_WARRIOR = 'parser/REMOVE_WARRIOR'
 export const SHOW_MESSAGES = 'parser/SHOW_MESSAGES'
 export const HIDE_MESSAGES = 'parser/HIDE_MESSAGES'
-export const ADD_NOTIFICATION = 'parser/ADD_NOTIFICATION'
-export const REMOVE_NOTIFICATION = 'parser/REMOVE_NOTIFICATION'
 
 export const PARSE_REQUESTED = 'parser/PARSE_REQUESTED'
 export const ADD_WARRIOR_REQUESTED = 'parser/ADD_WARRIOR_REQUESTED'
